Single Socket Client
Smoltcp does not dynamically generate sockets to keep a listener
open on the port. So the YCSB Client might find the server socket
busy closing the previouse connection and will terminate the request.

To avoid this we use one permanent tcp stream among client and server.
This in turn hasthe downside, that requests are no loger separated at
tcp level, and while we have no appropriate parsing in place at the
server and since we actually only want to benchmark the network layer
we just answer everything with "OK" for now.